requirements.txt
fastapi
uvicorn[standard]
azure-cognitiveservices-speech
openai